Tables_deleteOneRow and Tables_deleteOneColumn
These functions choose a table row or column within or adjecent to the
selection and delete it. The intention is for them to be called when the
user presses the "-" buttons on the table details sub-toolbar.

Both functions choose empty rows/columns in preference to those that
have content. They first look at the row/column that is directly below
or to the right of the selection, and if it's empty, delete that.
Otherwise, they look at the row/column directly above or to the left of
the selection, and if it's empty, delete that. Otherwise, they chose the
bottom/right row/column of the selection and delete that.