Tables: Make +/- buttons change rows/columns
Respond to changes in the rol/column steppers in the table details
sub-toolbar, by inserting or removing rows/columns from the table.

We do this by using KVO to monitor the rows and cols properties of
TableDetailView, requesting that the KVO mechanism provide both the old
and new value to us. If we see an increase, we do an insert; if we see a
decrease, we do a deletion. The latter uses the Tables_deleteOneRow and
Tables_deleteOneColumn functions added in the previous commit.

This provides a much more intuitive mechanism for adjusting the table
dimensions than the buttons surrounding the selection box, as it can be
used at any time the cursor or selection is within a table; unlike the
old mechanism, it does not require there to be multiple cells selected.
It is also useful for adjusting the dimensions right after you have
inserted a table, and is consistent with the way in which the dimensions
are specified prior to insertion.