Tables: Rename add row/column functions
Previously, we had four functions for adding rows and columns: one for
each edge of the selection box. Now that we're no longer providing +
buttons on the selection box for this purpose, and instead relying on
the -/+ buttons on the row/column steppers in the table sub-toolbar, we
now want functions that are designed & named for that purpose.

This commit renames/removes the following functions:

Tables_insertRowAbove     ->  removed
Tables_insertRowBelow     ->  Tables_addAdjacentRow
Tables_insertColumnLeft   ->  removed
Tables_insertColumnRight  ->  Tables_addAdjacentCol

All test files relating to these functions have also been renamed or
removed as appropriate.