Exclude paragraph style in lastReceivedFormatting
This is stored separately in EDEditor's paragraphStyleId property, and
is not considered part of the CSS properties for the selection. We make
this change to avoid passing a null -uxwrite-paragraph-style property to
the javascript applyFormattingChanges function.

There isn't any incorrect behaviour fixed by this change, but it just
seems logical to exclude this information from what we send to
applyFormattingChanges.