iOS 8: Temporary hack for input granularities
Our implementation of the UITextInputTokenizer protocol is incomplete in
that it does not handle some granularities, such as "sentence". This was
not used by iOS 7 but is in iOS 8 (by the system keyboard, and possibly
third-party keyboards as well). In fact a call to one of the input
methods with "sentence" granularity was made when a document was first
open, meaning the user would be presented with a javascript error
immediately after opening any document.

This simply gets around the problem by returning a very approximate
value, e.g. false if it's asking whether the position is at a sentence
boundary, and treating sentence granularity in the same way as paragraph
granularity in other cases. Ultimately we need to implement this
properly, by making sure we always return the same results as the
built-in UITextInputStringTokenizer class for a given piece of text.