util.js: Prefix all exports with Util_
In preparation for moving the codebase towards a proper module system,
ensure that all functions defined in util.js are prefixed
appropriately. This is necessary so we can easily identify all cases
where functions from the module are used, and set up the appropriate
dependencies for other modules that use them.

In util.js itself, we also wrap all the definitions in an anonymous
function, in order to give us a private scope in which we can define
functions and variables used only within the module. We don't actually
take advantage of that ability in this instance; but we're likely to do
so in the future, and this brings the code structure in line with other
files.