Rename Position class to Position_Position
Currently the codebase uses the symbol 'Position' for two purposes:
firstly, to provide a namespace for a set of functions that operate on
positions, and secondly, for a class of that name.

With the upcoming change to explicit referencing of modules as symbols
in their own right, we need to disambiguate between the two. For the
time being, the simplest solution seems to be just to add the prefix to
the class name, so that it's clear what namespace it resides in. We can
revisit this after the transition to TypeScript is complete.