Rename NodeSet.js to Collections.js
This file contains both the NodeSet and NodeMap classes. We want to put
these in a module that matches the filename, so to avoid having to use
NodeSet_NodeSet and NodeSet_NodeMap (which would be the case if we kept
the same filename), let's use the prefix Collections_ for both classes.

This commit just renames the file; the next commit will rename the
classes themselves (git is better at tracking file renaming when the
content doesn't change).