Use objects for modules, not global variables
Up until now, every function defined by every module was referenced via
a global variable, of the form ModuleName_functionName. Best practice in
modern javascript development dictates that code should be split into
separate namespaces/modules more cleanly, so that the dependencies
between files can be made more explicit.

This commit, which touches practically every file, changes the way in
which functions are accessed by creating an object for each module, and
storing references to the functions as properties of the object. This
moves us closer to being able to use a require()-style syntax as used in
CommonJS, or alternatively, a define()-style syntax as used in AMD. See
the folling for an explanation of these two styles:

    http://requirejs.org/docs/whyamd.html

The main difference in the majority of code is that that instead of
having a '_' character between the module name and function, you will
now see a '.' character. For example,

    Cursor.insertCharacter('X')

is now:

    Cursor_insertCharacter('X')

For defining modules, we previously had a list of global variables at
the top of the file, followed by an immediately-invoked function
expression that would set those variables, e.g.:

    var Cursor_insertCharacter;

    (function() {

        Cursor_insertCharacter = function(c) {
            // ...
        }

    })();

Instead, we now always refer to functions from the modules via an 'api'
variable, which is passed into functions like the above. The calls to
these use the special 'globalAPI' object (defined in first.js), which is
not supposed to be accessed from within normal code, and will be further
hidden in the future. We also have a series of variable declarations at
the top of the function scope, which play the same role as require()
calls in CommonJS, but work with our stil-in-progress API structure. So
the above code now looks like this:

    (function(api) {

        var Cursor = api.Cursor; // export

        var Clipboard = api.Clipboard; // import
        var DOM = api.DOM; // import
        // ... more imports

        Cursor.insertCharacter = function(c) {
            // ...
        }

    })(globalAPI);

A future step to bring us into compliance with the CommonJS/TypeScript
way of writing modules will be to move all of the code out of function
scope, replace the top-most variable declarations with calls to
require(), and in the case of TypeScript, use the 'export' keyword on
all appropriate function definitions.

The above module style is used for the .js files in src, as well as all
the .js files used by the test cases. For the test cases themselves, the
performTest() function now takes the 'api' object as a parameter, and
accesses all functions through that, instead of the no-longer-existing
global names. For example, the javascript code in the first test case
for insertCharacter now looks like this:

    function performTest(api)
    {
        api.Cursor.insertCharacter("X");
        api.tests.TestLib.showSelection();
    }