Set up test module objects in first.js
Prior to loading the javascript files used in the tests, ensure that an
object is created for each module. This removes the need to run the
files in a specific order, since each is able to get access to the
modules it needs (by assigning to a var, to simulate import), regardless
of whether the dependent module has been initialised yet.