Use AMD-style constructs for defining modules
Instead of directly modifying the shared 'api' object, use a define()
function which accepts parameters specifying the name of the module and
a function that, when executed, populates the contents of the module by
assigning functions to properties of the exports variable.  For the
purposes of all .js files (but not JS code in the .html test files),
this 'api' object is completely hidden.

Currently, we have a custom define() function which provides the minimum
functionality necessary for our needs. It is, however, compatible with
its counterpart in require.js. Our implementation of define() modifies
the globalAPI object, which is passed as the 'api' parameter to all the
test cases, since these rely on being able to refer to modules via
property access on the object, without using require() calls.

The start of a module definition now looks like this:

    define("Cursor",function(require,exports) {
        var Clipboard = require("Clipboard");
        var DOM = require("DOM");
        // ...
    }

The module's function populates the exports object with one property for
each export (most of which are functions, but a few of which are other
objects). All modules have been modified to use a named function syntax,
filling in the exports at the end of the file, e.g.:

    function ensureCursorVisible(center) {
        // ...
    }

    // ... more function definitions

    exports.ensureCursorVisible = ensureCursorVisible;
    // ...

Because there are many circular dependencies between our modules, we
cannot rely on the object return by a given require() call to be
populated immediately. So all cross-module function calls are made via
property-acecss syntax, e.g. DOM.nodeOffset(node).