Use .getTime() when getting interval between dates
TypeScript complains about our attempt to subtract date objects. What we
really want (and in practice get with JS) is the number of milliseconds
between the two Date objects. To fix the compile error, call the
getTime() method, which returns the number of milliseconds since the
epoch, before doing the subtraction.