Remove one level of indentation
Up until now, all functions defined within modules have had an initial
indentation of four spaces at the start of each line, due to the fact
that they exist inside a wrapper function for the purposes of
establishing a module-local scope, and invoking AMD's define() function.

With the transition to TypeScript, we will be using a CommonJS-like
syntax, in which the wrapper function is unnecessary. To keep the diffs
easier to look at past the jump to TypeScript, the indentation of all
code has been decreased one level, so we won't have to change it later.

This does of course have the downside of making diffs between versions
of files prior to this commit and those afterward difficult to read,
although passing the -w option (ignore whitespace) to git diff can be
used when such a need arises.