Mark function args as optional where appropriate
Fix all instances of the error "Supplied parameters do not match any
signature of call target". In almost all cases, this is because some
function calls did not include all of the arguments given in the
function declaration, but those arguments were always intended to be
optional. In pure JS this is not a problem, but TS requires arguments to
be explicitly marked as optional by appending a ? to the name.

There were also a couple of cases where a function call was passing
unnecessary extra arguments where it should not have been.