Use 'any' type for custom objects
Fix all instances of the error "Property 'X' does not exist on type
'Object'". These occurred when we would do something like the following:

    var result = new Object();
    result.start = // ...
    result.end = // ...
    return result;

This error occurs because TypeScript knows that the Object class does
not have these properties defined. The (temporary) solution is to add a
type annotation to the variable to mark it as type 'any', which will
allow us to set whatever properties we want. Ultimately, we should be
using a class or interface in these places.