Use TypeScript import syntax
Use TypeScript's import and export statements for defining the
dependencies and public functions/variables for all modules. This
replaces the previous arrangement where we explicitly used our own
AMD-like API, which involved calling define() for each module, and
setting properties on an 'exports' variable.

For example, instead of this:

    define("Traversal",function(require,exports) {
        var Util = require("Util");
        function foo() {
            Util.bar();
        }
        exports.foo = foo;
    });

We now do this:

    import Util = require("./util");
    export function foo() {
        Util.bar();
    }

Because the "module" setting in tsconfig.json is set to "amd", the .js
files produced by the TypeScript compiler are essentially the same as
what we were doing manually before. The only differences are:

1) Module names are the same as the filename (including the path), minus
the .js extension. They are derived automatically from the filename, not
specified explicitly in the call to define().

2) The argument to require() is now the path to the file, minus the
extension. The path is relative to the directory containing the file
doing the import.

Because modules are now identified by paths, significant changes also
had to be made to our module loader (in first.js) to accommodate this,
while retaining backwards compatibility with the old method. That uses a
global api object with properties based on the old module names, which
is still relied upon by the tests. None of the tests use AMD to load
their dependencies but instead accept an "api" argument to
performTest(), on which they do things like api.UndoManager.newGroup().
So the module loader sets this up, but maps the filenames, which differ
slightly from the old module names, to the appropriate properties.