Use instanceof instead of .nodeType
Many places in the code check to see if a node is a text node or an
element, and then access properties specific to one or the another.
Until now, we've done this by checking the nodeType property of the
object against Node.TEXT_NODE or Node.ELEMENT_NODE. However, if we add
type annotations indicating the variable is an instance of Node, then
attempts to access properties only present on subclasses cause compile
errors, in the absence of explicit casting.

TypeScript has a feature called "type guards", where it can identify
conditional statements that check the type of an object, and then within
the relevant block, consider the variable to have that type. However,
this does not pick up 'if (x.nodeType == Node.TEXT_NODE)'. Instead, we
now say 'if (x instanceof Text)', which the compiler recognises.

Right now this doesn't make any practical difference, but it will make
the imminent process of adding type annotations in certain parts of the
code easier.