Fix parameters in function types
There were a few places where I'd used => to define a function type, in
a formal parameter to another function. I had incorrectly assumed that
you could write something like:

    (number) => string

and that would mean the parameter must be a function which takes a
single number parameter, and returns a string. However, the above
actually just specifies the *name* of the parameter, not its type.
Instead, it is necessary to specify both:

    (n: number) => string

Note that the name given in the type signature doesn't have to match the
actual name used by a function conforming to the type. Strictly
speaking, the type system does not need the name - it's just required by
the syntax.