Avoid resource loading warnings in tests
Many tests still included references to external .js files that are now
part of the typescript code in tests/ and no longer exists at their
original locations. These were causing warnings to be printed to the
console during test runs.

A number of tests also insert images into the documents, where the image
file referenced does not exist. The dummy image used for some of the
other tests has now been copied to the relevant directories. As a
result, there should be no warnings visible on the console after a test
run now.