Always use numeric ids for input tests
All of the public Input APIs use numeric identifiers to represent
position objects, to make it simple to access them from other languages.
In the tests however, we would sometimes pass a Position object into an
Input function, which in turn would pass that through to getPosition().

In order to be able to give getPosition() a type for its input
parameter, we now only ever pass in numeric identifiers from the tests -
calling Input.addPosition() to generate one for a Position object where
necessary.