Make DoublyLinkedList a class
I'd put this off until now because it was a slightly tricky case of
typing - we need to use generics to specify the type of items in the
list, and those items must have 'prev' and 'next' properties that are of
the same type.

This was achieved by adding a generic DoublyLinkedListItem<T> interface,
which DoublyLinkedList requires that its item type comply with. Note
that the type constraint is:

    <T extends DoublyLinkedListItem<any>>

and not:

    <T extends DoublyLinkedListItem<T>>

because otherwise the compiler gives the errorr "TS2313: Constraint of a
type parameter cannot reference any type parameter from the same type
parameter list". For a explanation of the workaround needed here, see
http://stackoverflow.com/questions/23003188.

Note that because TypeScript uses structural rather than nominal typing
(that is, it compares types based on their structures, rather than their
names), it is not necessary for the item type of a DoublyLinkedList to
explicitly declare that it implements the DoublyLinkedListItem
interface. For example, the OutlineItem class in outline.ts did not need
to be changed, because it already complies with the necessary
constraints of the DoublyLinkedListItem type (that is, it has 'prev' and
'next' properties which point to objects of the same type).