Make Util independent of all other modules
There were only two functions in the Util module which needed code from
other modules: debug(), which called directly through to Editor.debug(),
and nodeHasContent(), which used the Types and ElementTypes modules.

This commit replaces all calls to Util.debug() to Callbacks.debug(), and
moves the nodeHasContent() function into the Types module.

Additionally, there were only two places in Types which relied on DOM -
calls to getAttribute() - and these are now replaced with direct calls
to the equivalent method on the node object itself. Thus, both Util and
Types are now independent of the DOM module.