Make Traversal independent of Callbacks
The only reason Callbacks was used in Traversal was for the printTree()
function, which is only for debugging. This has now been replaced with
treeToString(), which is actually more useful now anyway because the
caller can decide what to do with the output, instead of just having it
printed to the console. The function isn't actually used anywhere at the
moment anyway.