Make Position independent of DOM
To avoid circular dependencies, and eventually allow us to expose just
the Position class from position.ts (which we'll do in a future commit),
remove the dependency on the DOM module, which itself references the
class.

The main change to be made is in the way we store the array of tracked
positions for a given node (i.e. the positions that should be modified
if the node is moved or deleted). Previously, we had a variable in the
DOM module called nodeData, which was a map of node ids to NodeData
objects, each of which in turn contained a trackedPositions array.
However, this data was private to the DOM module, and could not be
accessed directly from the Position class - which would instead use the
addTrackedPosition() and removeTrackedPosition() functions to modify it.

Now, we remove the NodeData class, and use the _trackedPositions
property of the Node object itself to store the array. This can be
accessed directly from both modules, removing the need for a dependency
on the former for the purpose of tracking.

The only other DOM functions used in Position are nodeOffset() and
maxChildOffset(). These now live in the Traversal module, though
wrappers remain in the former because they are still referenced there by
many other parts of the code.