Remove workaround for iOS 7 getClientRects() bug
There was a bug in iOS 7 which caused getClientRects() on the built-in
DOM Range class to return coordinates relative to top of document, when
it should instead return coordinates relative to the current client view
(that is, taking into account scroll offsets). Since this was fixed in
iOS 8, we no longer need the workaround.

More importantly, this allows us to remove the dependency that the Range
module had on Main, since the variable indicating whether the workaround
should be enabled was stored in the latter. Main.init() has also been
modified so it no longer takes the flag as a parameter.