Support direct single exports in module loader
Make it possible to export a single value, object, or class from a
module, instead of a using a pre-supplied exports object that is
populated with one or more properties.

To achieve this, we perform a topological sort of the dependency graph
and then load the modules in the appropriate order. This ensures that a
given module completes initialisation before any others that depend on
it.

For modules that populate the exports object, circular references are
still supported as before. For modules that return a single value,
circular references are now detected by the loader, which throws an
exception if any are found.