Don't use a global object for exposing modules
For the test harness, we were relying on the presence of a globalAPI
object which had references to all of the modules that had been loaded.
To get better encapsulation and more closely comply with the AMD module
loading approach, hide all module references within the loader, and only
make them accessible from within other modules that explicitly import
them.

The *only* way to get a reference to a module now is via a call to
define(), including the relative module path in the dependency list.
This is the existing AMD-compatible approach we've been using in all
cases except for the test harness, and is what the TypeScript compiler
is configured to output in tsconfig.json.

To solve the problem of how to access the test code loaded into the
iframe used by the test harness, we have a new module, testrun.js, which
exposes a global function run() on the iframe's window object. The test
harness calls this function, which then calls performTest() in the HTML
file that has been loaded for a given test.

The tests still rely on the "all modules accessible via an object"
approach as was previously the case, but this object is now created in
testrun.js itself, conforming to the naming scheme that the tests
assume. As before, performTest() in each test file takes this object as
a parameter.

The module loader no longer needs to maintain the specific object/
property arrangement used by the tests, and is free to store its
collection of modules however it chooses. Presently, this is the
existing modules variable defined within a private scope; globalAPI is
gone.