Pass name parameter to define() in test harness
Each .js file produced by the TypeScript compiler contains a single call
to define(), which registers the corresponding module. The module loader
needs a way to determine the name of the module, one of which is to have
the name supplied as the first parameter. The generated code does not do
this however, as the compiler assumes that the loader will determine the
module name from the script filename - as is the case with fully-
featured AMD loaders like require.js.

In our test harness however, we execute the module code via eval(),
instead of injecting <script> elements into the DOM. The main reason for
this is efficiency; we load all test files and concatenate them together
into a single string kept in memory for the duration of a complete run
of all tests, calling eval() on the string for each individual test.
Because of this, there is no direct way for the loader to determine
which file each modules was defined in.

We used to get around this by having a _nextDefineFilename property on
the window object, which define() would inspect. Upon loading each file,
prior to concatenation, we would add a line of code which sets this.
However, this is a non-standard approach and is not compatible with the
way in which other AMD module loaders work.

Instead, the readModule() function in testharness.ts now modifies the
code it reads from the file, inserting the name parameter to the
define() call using string substitution.  This means that define() does
not need any special tricks to determine the module name.