Support for <script> elements in module loader
In the module loader, if the define() function is called without giving
a name for the module, determine it from the filename of the script
element that contains the define() call.

This currently relies on the assumption that all scripts are loaded
synchronously, and that the <script> element being executed is the one
that appears last in document order. It may be possible to relax this
restriction in the future.

This commit also adds a version of the demo file which works by
injecting script elements into the page and using our own module loader,
in contrast to the existing demo.html which uses require.js.