Make require() available after loadAllModules()
Allow JavaScript code to call the require() function once all modules
have been defined and loaded.  This makes it unnecessary to use tricks
like exporting global variables from within certain modules in order to
access them from the "outside" (e.g. JS code in the HTML page), as we
did previously for example in demo-customloader.html. Instead, a call to
require() is made, passing in the name of the desired modules, and a
function whose arguments will be filled in with the modules.

As part of this change, the loader is no longer restricted to files in
the build directory. It allows the HTML page to supply a base URL from
which to resolve the relative module paths passed to require(), and this
is used to determine the absolute name of the module.