Make require() usable if no mainScriptURL is set
The require() function is made available after loadAllModules() has
completed, and is intended for use in JavaScript code embedded directly
within a HTML file. In order for module names to be specified as
relative paths, it needs to know the base URL by which these should be
resolved. It uses this in conjunction with window.location.href to
determine the absolute name of a module.

However, ther are cases where the we do *not* want the resolution to be
influenced by the URL of the current page, such as when an arbitrary
HTML file is loaded for editing, and the library code is injected by
eval() or equivalent (e.g. stringByEvaluatingJavaScriptFromString on
iOS). With this commit, require() now treats the supplied module names
as absolute if loadAllModules() is passed a null base URL, which is what
you should do if you are injecting code to a page as just described.