Use shorter names for functions in input module
The functions in the input module exist to support iOS's UITextInput and
UITextInputTokenizer protocols, both of which have several methods with
really long names, due to Objective C's insistence on including the
names of all parameters after the first as part of the method signature.
I'd originally translated all the method names directly, but with this
commit we use abbreviated versions.