Define all types used by external API in one file
Put the type definitions of all interfaces and classes used by the
external API into externalapi.ts, and import those types from the
implementation modules that use them, rather than the other way round.

The purpose of this is to make the definition file produced by the
compiler, externalapi.d.ts, all-inclusive. All types used by API
functions are defined in this same file, and it does not contain any
import statements. This gives us a completely-encapsulated API which
means that third-party code using the library is not made aware of any
details about the implementation modules.

This commit also turns on the option for definition file generation in
tsconfig.json, and adds the build directory to the list of paths that
the compiler should ignore when searching for input files. Without the
latter, on the second and subsequent runs of the compiler, it complains
about conflicts with types being re-defined because they are present in
both the .d.ts files in the build directory (produced on the first run),
and the .ts files in the src directory.