Add PositionRef and RangeRef
The input module in the external API references Position objects using
opaque values, to avoid exposing the Position objects used internally to
the outside world. This ensures that all uses of Position objects are
safe, e.g. it is not possible for application code to create a position
with a null node or an invalid offset.

Up until now, we've been using numbers for the opaque references. This
is not ideal, as a generic type like this does not convey any indication
that it specifically means a reference to a Position object. Indeed, it
is conceivable that some developers using the API could incorrectly try
to use the numbers for other purposes, e.g. by attempting to perform
arithmetic operations on them, which would make no sense.

In this commit, we introduce the PositionRef interface, which is used
for all parameters and return values in the input module that refer to
positions, and were previously typed as number. The type makes it clear
what the value means.

The PositionRef interface has a single string field named PositionRefId.
TypeScript's use of a structural rather than a nominal type system means
that type checker considers only the names and types of the fields of an
interface when determining compatibility, ignoring the name. By using a
special field name that's likely to be unique to this interface, we can
sort of "fake" nominal typing.

While nominal typing would be preferable in this case for applications
written in TypeScript, the structural typing also has the benefit of
enabling the opaque reference objects to be easily exchanged across
language boundaries, since they can be encoded as JSON, e.g.:

    { "PositionRefId": "some-value" }

Applications written in other languages which invoke the API from within
an embedded web view can thus receive and supply the reference objects
in the above format. Applications written in TypeScript get the benefit
of compile-time type checking by virtue of the compiler only accepting
types that have the same structure as PositionRef, which in practice is
going to be only those objects returned methods in the input module.

Finally, this commit also introduces the RangeRef interface, which is
simply a pair of PositionRef objects, named start and end. The types of
RangeRef variables be checked by the compiler in the same manner as for
PositionRef, and they can also be encoded as JSON.