Notify application when document is modified
After a call to any API function which may cause the document to be
modified, send a "documentModified" callback message to indicate that
the version of the document in memory differs from that which was
originally loaded. Applications that respond to this callback message
can then mark the document as "dirty", handling this e.g. by prompting
the user if they want to save the document when they go to close it.

Previously, the only way for achieve this same was for the application
to maintain a list of all the API functions and maintain a mapping of
which functions modified the document content (e.g. inserting/deleting
text) and which ones didn't (e.g. moving the cursor or adjusting the
selection). This distinction is now codified in the use of two different
methods to wrap APIs - execute() and modify() in externalapi.ts. The
latter sends the "documentModified" callback message, while the former
doesn't.