Prefix position ids differently for anchor/dynamic
Position references can be "anchors" - which always correspond to a
fixed concept (document/selection start/end), or "dynamic" positions -
which are generated at runtime and can represent arbitrary positions
within the document.

To simplify the logic of mapping between references and concrete
Position objects, and to prepare for upcoming changes to the retain/
release logic for positions, prefix each reference id with either an "A"
or a "D" to indicate whether it's an anchor or dynamic position.