Invalidate position refs on document modification
Change the rule dictating the lifetime of PositionRefs so that instead
of them remaining valid until removePosition() is called, they now
remain valid until the document content is modified - which is defined
by any call to an API that uses the modify() function. The latter
automatically invalidates all PositionRef objects after the relevant API
call has completed, which means that the Position objects themselves are
no longer tracked, and the map of reference ids to Position objects
maintained in the inputref module is cleared. The application receives a
"documentModified" callback message when this happens, so it can
appropriately clear or mark as invalid any data it maintains that
relates to previously-obtained positions.

I don't currently see a need for PositionRefs to remain valid after
document modification; the purpose of these objects, and of the input
module in general, is to answer questions about the content of the
document for the purposes of cursor movement and selection.

If necessary, we could add a reference counting mechanism to the API so
that the application can indicate it wants a PositionRef to remain valid
for as long as it still needs it. This is what I assumed was required by
the iOS UITextInput and UITextInputTokenizer protocols (which this
module was originally designed for); however it turns out that the iOS
text input system does not use positions after modification anyway, so
we don't need explicit invalidation from the application in that
particular case at least.