Add the EventReceiver interface
Up until now, the only way to retrieve information about events (like
cursor movement or outline changes) that had occurred during an API call
was by invoking the getBackMessages() function, which return a JSON-
encoded array of messages. This is fine for bindings to other languages
(e.g. Objective C) where the binding has an opportunity to do stuff
(like invoke callbacks) before returning the value of the API call to
the application. It is not appropriate, however, for JavaScript or
TypeScript code that is using the API directly.

Now, by defining a class which implements the EventReceiver interface,
an application can supply an instance of this to the API via the
setReceiver() function. Every API call will dispatch any event messages
to the object right at the end of its execution, by invoking the
relevant methods on the object. The old approach of storing the message
data is still available for applications which wish to use that
(currently, UX Write).

Note that we still maintain the restriction that the event callbacks are
not invoked at the time at which the events are posted, since it's
possible that the callbacks may go on to invoke further API functions,
and none the APIs exposed by the library are guaranteed to be
re-entrant.