Display cursor directly from within editor
Instead of requiring the host application to draw the cursor (triggered
by the "setCursor" event), display it by creating a div in the selection
module's update() function, in a similar manner to how we display
selection highlights and the borders of table selections.

The host application can use the setCursorHidden() function to disable
this functionality if desired, or to flash the cursor on/off using its
own timer mechanism. We avoid using a timer within the library, to
ensure that the application has control over exactly when JavaScript
code runs.