Make cursor/selection via mouse work in Firefox
WebKit, Blink, and Edge all implement the caretRangeFromPoint() function
which we've been using since the beginning to determine the position in
the DOM tree corresponding to a particular set of coordinates. Firefox
instead provides a different function, caretPositionFromPoint(), that
does essentially the same thing. We now check which of these two
functions is present, and use the appropriate one.

Unfortunately IE (including IE11) supports neither, so if we wish to
support IE, we'll need to come up with some other way of getting this
information.