Fix document-length cursor at left of screen
When a mouse click event occurs and we try to determine the position in
the document where the cursor should go, we use positionAtPoint(). This
queries the browser to determine the node and offset corresponding to
the specified coordinates.

This normally worked fine, except if you clicked on a div element
representing the cursor, as created by the update() function in the
selection module. Because this covered the text on which you were trying
to click, the returned position would be of the div element, not the
text underneath it. The cursor is only 2 pixels wide, so this effect
only happened if you clicked in exactly the right spot. The outcome of
this bug was that the cursor would appear on the far left of the screen,
spanning the entire length of the document.

Now, we temporarily set the visibility property of the cursor div to
"hidden" while computing the position, ensuring we get the right result.
Because the browser only updates the screen after JavaScript execution
completes, the fact that we unhide thes immediately afterwards does not
result in the temporary hiding being apparent to the user.